+++ lib/boxcars/prompt.rb
@@ -19,11 +19,10 @@ module Boxcars
     # compute the prompt parameters with input substitutions (used for chatGPT)
     # @param inputs [Hash] The inputs to use for the prompt.
     # @return [Hash] The formatted prompt { messages: ...}
-    # rubocop:disable Lint/UnusedMethodArgument
+    # rubocop:disable-next Lint/UnusedMethodArgument
     def as_prompt(inputs: nil, prefixes: nil, show_roles: nil)
       { prompt: format(inputs) }
     end
-    # rubocop:enable Lint/UnusedMethodArgument
 
     # compute the prompt parameters with input substitutions
     # @param inputs [Hash] The inputs to use for the prompt.